requirements.txt
streamlit
google-generativeai
pymupdf
faiss-cpu
numpy
streamlit-chat